feat: Destructure Supabase client into multiple files
Refactors the monolithic `supabase_client.py` into smaller, semantically organized files based on the database table they interact with.

- Creates new client files for actions, documents, mails, requests, and tasks.
- Moves relevant functions, constants, and Pydantic models to their respective new files.
- The core `supabase_client.py` now only contains common client setup and authentication functions.
- Updates the import statement in `bucket_client.py` for clarity.

This change improves code organization and maintainability without altering functionality.
